Added bitwise convolution transpose implementation
