+++ bitwise_network.py
@@ -63,8 +63,8 @@ class BitwiseNetwork(nn.Module):
         # Initialize inverse of front end transform
         self.scale = kernel_size / stride
         inv_basis = torch.t(torch.pinverse(self.scale*basis))
-        self.conv1_transpose = BitwiseConv1d(transform_channels, in_channels,
-            kernel_size, stride=stride, biased=False)
+        self.conv1_transpose = BitwiseConvTranspose1d(transform_channels,
+            in_channels, kernel_size, stride=stride, biased=False)
         self.conv1_transpose.weight = nn.Parameter(inv_basis.unsqueeze(1),
             requires_grad=adapt)
 
@@ -210,80 +210,6 @@ def val(model, dl, loss=F.mse_loss, device=torch.device('cpu'), autoencode=False
         metrics = bss_eval_batch(estimates, sources)
         bss_metrics.extend(metrics)
     return running_loss / len(dl), bss_metrics
-
-class LossMetrics(nn.Module):
-    '''
-    Data struct that keeps track of all losses and metrics during the training process
-    '''
-    def __init__(self):
-        self.time = []
-        self.train_loss = []
-        self.val_loss = []
-        self.sdrs = []
-        self.sirs = []
-        self.sars = []
-
-    def update(self, train_loss, val_loss, sdr, sir, sar, output_period=1):
-        if self.time:
-            self.time.append(self.time[-1] + output_period)
-        else:
-            self.time = [0]
-        self.train_loss.append(train_loss)
-        self.val_loss.append(val_loss)
-        self.sdrs.append(sdr)
-        self.sirs.append(sir)
-        self.sars.append(sar)
-
-def train_plot(vis, loss_metrics, eid=None, win=[None, None]):
-    '''
-    Plots loss and metrics during the training process
-    '''
-    # Loss plots
-    data1 = [
-        dict(
-            x=loss_metrics.time, y=loss_metrics.train_loss, name='Training Loss',
-            hoverinfo='y', line=dict(width=1), mode='lines', type='scatter'),
-        dict(
-            x=loss_metrics.time, y=loss_metrics.val_loss, name='Validation Loss',
-            hoverinfo='y', line=dict( width=1), mode='lines', type='scatter')
-    ]
-    layout1 = dict(
-        showlegend=True,
-        legend=dict(orientation='h', y=1.1, bgcolor='rgba(0,0,0,0)'),
-        margin=dict(r=30, b=40, l=50, t=50),
-        font=dict(family='Bell Gothic Std'),
-        xaxis=dict(autorange=True, title='Training Epochs'),
-        yaxis=dict(autorange=True, title='Loss'),
-        title=win[0]
-    )
-    vis._send(dict(data=data1, layout=layout1, win=win[0], eid=eid))
-
-    # BSS_EVAL plots
-    data2 = [
-        # SDR
-        dict(
-            x=loss_metrics.time, y=loss_metrics.sdrs, name='SDR',
-            hoverinfo='name+y+lines', line=dict( width=1), mode='lines', type='scatter'),
-        # SIR
-        dict(
-            x=loss_metrics.time, y=loss_metrics.sirs, name='SIR',
-            hoverinfo='name+y+lines', line=dict( width=1), mode='lines', type='scatter'),
-        # SAR
-        dict(
-            x=loss_metrics.time, y=loss_metrics.sars, name='SAR',
-            hoverinfo='name+y+lines', line=dict( width=1), mode='lines', type='scatter'),
-    ]
-    layout2 = dict(
-        showlegend=True,
-        legend=dict(orientation='h', y=1.05, bgcolor='rgba(0,0,0,0)'),
-        margin=dict(r=30, b=40, l=50, t=50),
-        font=dict(family='Bell Gothic Std'),
-        xaxis=dict(autorange=True, title='Training samples'),
-        yaxis=dict(autorange=True, title='dB'),
-        yaxis2=dict(autorange=True, title='STOI', overlaying='y', side='right'),
-        title=win[1]
-    )
-    vis._send(dict(data=data2, layout=layout2, win=win[1], eid=eid))
 
 def main():
     parser = argparse.ArgumentParser(description='bitwise network')
